IOS-3217 Add date formatter + rename section to group
