+++ Anytype/Sources/PresentationLayer/TextEditor/EditorPage/Views/Settings/History/VersionHistoryDataGroup.swift
@@ -1,0 +1,17 @@
+import Foundation
+import Services
+
+struct VersionHistoryDataGroup: Identifiable, Hashable {
+    let title: String
+    let icons: [ObjectIcon]
+    let searchData: [VersionHistoryItem]
+    
+    var id: Int { hashValue }
+}
+
+struct VersionHistoryItem: Identifiable, Hashable {
+    let id: String
+    let time: String
+    let author: String
+    let icon: ObjectIcon?
+}
