--- Anytype/Sources/PresentationLayer/TextEditor/EditorPage/Views/Settings/History/VersionHistoryDataSection.swift
@@ -1,9 +1,0 @@
-import Foundation
-import Services
-
-struct VersionHistoryItem: Identifiable, Hashable {
-    let id: String
-    let time: String
-    let author: String
-    let icon: ObjectIcon?
-}
